MQ integration for #41
* Added an /api/droplet endpoint that is only accessible via an admin
api_key. API keys are random hashed strings generated when a user is
created.
* Added the swiftmq plugin that publishes channel deletions and
additions to a topic exchange on the MQ.
* Moved full url extraction from Model_Link since the link extractor
connected to the MQ does this already. The local link extractor plugin
should do this job.
* Added semantics_complete and links_complete that is used by the
api/droplet endpoint to determine whether processing of a droplet is
complete and make it available to the user.