River expiry. Closes #104
* Rivers now expire after 14 days. Each time, the river is accessed we
check if it has expired and we surface an alert message with a link to
extend the lifetime of the river. Clicking on the link extends the
lifetime of the river by the default lifetime period. There is also an
additional river expiry check prior to adding droplets to a river -
This is in the add_metadata() method of the Model_Droplet class

* The rivers table has 3 new columns - "river_date_expiry",
"river_expired" and "lifetime_extension_token". The latter maintains a
token that is used to authenticate a request to extend the lifetime of
a river while the other two columns track the expiry date of the river
and whether the expiry status of the river.

* Removed on-the-fly update of the site settings in the "application"
tab. The settings update must now be triggered by clicking on the "Save
Changes" button.