Update installation instructions to match those in the wiki
* The previous instructions required users to download the code from
the GitHub repository. However, the tarball from the repo does not
include the contents of the directories referenced by the submodules.
This fix syncs the basic installation instructions with those on the
wiki - https://wiki.ushahidi.com/display/WIKI/Installing+SwiftRiver.
Closes #278 - largely an issue of incorrect instructions

* For release iteration, once the code has been tagged with the release
number, a tarball of the application will be uploaded to GitHub for
download