Move analytics to core
* Analytics should really be part of the core distribution
and not a plugin that is shipped separately because it is a 
feature that is beneficial to all users. I have therefore
merged the swiftriver-analytics-plugin into the SwiftRiver
repo.

* Deleted redundant unit tests and classes. The unit tests
were written when building out the initial prototype of the
application but they were not being updated as we iterated
on the product.