Split test data into multiple files
Not all tables are required for every test so this allows only the
required tables to be reset.

Also add a new define in the testing bootstrap.php for TESTING_MODE.
The application bootstrap.php will then set Database::$default =
'unittest' when this constant is defined. This way, a 'unittest'
database configuration can be set in the database.php config file for
use by the unit tests.